lara/common: move storage to TRX
