+++ src/tr1/meson.build
@@ -110,7 +110,6 @@ sources = [
   'game/gun/gun_rifle.c',
   'game/inventory.c',
   'game/inventory_ring/vars.c',
-  'game/lara/common.c',
   'game/objects/common.c',
   'game/objects/setup.c',
   'game/option/option.c',
