+++ src/tr2/meson.build
@@ -103,7 +103,6 @@ sources = [
   'game/gun/gun_rifle.c',
   'game/inventory.c',
   'game/inventory_ring/vars.c',
-  'game/lara/common.c',
   'game/lara/flare.c',
   'game/level.c',
   'game/objects/common.c',
